+++ app_main.py
@@ -1,5 +1,6 @@
 import base64
 import os
+from flask import Flask, send_from_directory
 
 import dash
 
@@ -10,21 +11,20 @@ LOGOS = [
     for fn in os.listdir('logos') if fn.endswith('.png')
 ]
 
+HDR_LOGO = base64.b64encode(open('icons/header-logo.png', 'rb').read())
+
 INFO_ICON = base64.b64encode(open('icons/information.png', 'rb').read())
+
+REPORT_IMG = base64.b64encode(open('assets/report_frontpage.png', 'rb').read())
 
 PLACEHOLDER = base64.b64encode(open('assets/placeholder.png', 'rb').read())
 
 APP_BG_COLOR = '#FFFFFF'
 
-# Initializes dash app
-app = dash.Dash(__name__)
+server = Flask(__name__)
 
-# Load css file
-external_css = [
-    'https://raw.githack.com/rl-institut/WAM/dev/static/foundation/css/app.css',
-]
-for css in external_css:
-    app.css.append_css({"external_url": css})
+# Initializes dash app
+app = dash.Dash(server=server)
 
 app.index_string = '''
     <!DOCTYPE html>
@@ -34,7 +34,6 @@ app.index_string = '''
             <meta charset="UTF-8">
             <meta description="NDC visualisation">
             <meta name="viewport" content="width=device-width, initial-scale=1.0">
-            {%favicon%}
             {%css%}
         </head>
 
@@ -52,3 +51,33 @@ app.index_string = '''
 app.title = 'NDC visualisation'
 
 app.config.suppress_callback_exceptions = True
+
+
+@server.route("/download/<path:path>")
+def download(path):
+    """Serve a file from the upload directory."""
+    return send_from_directory('data', path, as_attachment=True)
+
+
+@server.route('/favicon.ico')
+def favicon():
+    return send_from_directory('assets', 'favicon.ico', mimetype='image/vnd.microsoft.icon')
+
+
+stylesheets = ['app.css', 'base.css', 'intro.css']
+static_css_route = 'assets'
+
+
+@app.server.route('/static/<stylesheet>')
+def serve_stylesheet(stylesheet):
+    if stylesheet not in stylesheets:
+        raise Exception(
+            '"{}" is excluded from the allowed static files'.format(
+                stylesheet
+            )
+        )
+    return send_from_directory('assets', stylesheet)
+
+
+for stylesheet in stylesheets:
+    app.css.append_css({"external_url": "/static/{}".format(stylesheet)})