+++ CHANGELOG.md
@@ -16,7 +16,21 @@ Here is a template for new release sections
 ### Removed
 -
 ```
-## [Unreleased]
+## [2.1.0] 2020-01-20
+
+### Added
+- Tool citation and impressum (#96)
+- Favicon (#97)
+- cc0 License for output data (#98)
+
+### Changed
+- New banner (#96)
+- Rename South America to Central and South America (#95)
+- Rename Universal-Energy-Access to Universal-Electricity-Access (#95)
+- Colorset for plotting (#97)
+- Serve files explicitly with flask (#98)
+
+## [2.0.0] 2019-10-31
 
 ### Added